UHCI device discovery

+++ kernel/src/serial/com.h
@@ -3,8 +3,8 @@
 #include <lib/types.h>
 
 u8 inb(u16);
-u8 inw(u16);
-u8 inl(u16);
+u16 inw(u16);
+u32 inl(u16);
 
 void outb(u8 data, u16 port);
 void outw(u16 data, u16 port);
